Move toc into maple_markdown crate
