+++ crates/maple_core/Cargo.toml
@@ -26,7 +26,6 @@ tokio = { workspace = true, features = ["fs", "rt", "process", "macros", "rt-mul
 markdown = { workspace = true }
 once_cell = { workspace = true }
 parking_lot = { workspace = true }
-percent-encoding = { workspace = true }
 rayon = { workspace = true }
 regex = { workspace = true }
 rgb2ansi256 = { workspace = true }
