+++ crates/maple_markdown/Cargo.toml
@@ -11,10 +11,13 @@ publish.workspace = true
 [dependencies]
 axum = { version = "0.7.5", features = ["ws"] }
 axum-extra = { version = "0.9.3", features = ["typed-header"] }
-headers = "0.4.0"
 markdown = { workspace = true }
+once_cell = { workspace = true }
+percent-encoding = { workspace = true }
+regex = { workspace = true }
 serde_json = { workspace = true }
-tokio = { version = "1", features = ["full"] }
-tower-http = { version = "0.5.2", features = ["fs", "trace"] }
+tokio = { workspace = true, features = ["full"] }
 tracing = { workspace = true }
 webbrowser = { workspace = true }
+
+utils = { workspace = true }
