--- crates/markdown/js/index.html
@@ -1,61 +1,0 @@
-<!DOCTYPE html>
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/github-markdown-css/5.5.1/github-markdown.min.css" >
-    <style>
-      .markdown-body {
-        box-sizing: border-box;
-        min-width: 200px;
-        max-width: 980px;
-        margin: 0 auto;
-        padding: 45px;
-      }
-
-      @media (max-width: 767px) {
-        .markdown-body {
-          padding: 15px;
-        }
-      }
-    </style>
-    <title>Markdown Preview</title>
-</head>
-<body>
-    <article class="markdown-body" id="content"></article>
-
-    <script>
-        var previewWindow = document.getElementById('content');
-
-        var webSocketUrl = 'ws://' + window.location.host;
-
-        // Connect to the WebSocket server using the port number
-        var socket = new WebSocket(webSocketUrl);
-
-        // Handle WebSocket events...
-        socket.onmessage = function(event) {
-            const message = JSON.parse(event.data);
-            if (message.type === "update_content") {
-                document.getElementById('content').innerHTML = message.data;
-            } else if (message.type === "scroll") {
-                console.log(`Recv message: {message}`);
-                const scrollPercent = message.data;
-                // Calculate the absolute scroll position based on the percentage
-                const windowHeight = window.innerHeight;
-                const totalScrollHeight = document.documentElement.scrollHeight - windowHeight;
-                const absoluteScrollPosition = totalScrollHeight * (scrollPercent / 100);
-
-                // Scroll the page to the calculated absolute position
-                window.scrollTo(0, absoluteScrollPosition);
-            } else {
-                console.log(`Invalid message: {message}`)
-            }
-        }
-
-        socket.onclose = function(event) {
-            // Close the browser window.
-            window.open('', '_self', '');
-            window.close();
-        }
-    </script>
-</body>
-</html>
